Refactor BufferAPIHandler and related components
- Updated TSAutoBuffer, VectorAutoBuffer, and GraphAutoBuffer initialization to remove namespace prefixes for cleaner code.
- Changed timestamp field in TSStore DataPoint from `timestamp` to `timestamp_ms` for clarity.
- Modified buffer statistics to use atomic load for thread safety.
- Updated logging to use the new logger header.
- Refactored blob transfer handler to use the correct proto namespace for BlobChunk and related enums.
- Added build scripts for Visual Studio and Ninja, ensuring proper environment setup and build verification.
- Introduced process mining functions with JSON serialization/deserialization for patterns and configurations, including new AQL functions for process mining operations.

+++ src/cdc/changefeed_buffer.cpp
@@ -199,7 +199,7 @@ size_t ChangefeedBuffer::flushBuffer(Changefeed::ChangeEventType event_type, Eve
     }
     
     auto span = Tracer::startSpan("ChangefeedBuffer.flushBuffer");
-    span.setAttribute("event_type", static_cast<int>(event_type));
+    span.setAttribute("event_type", static_cast<int64_t>(event_type));
     span.setAttribute("events", static_cast<int64_t>(buffer.events.size()));
     
     // Record all events to changefeed
